docs/proposals/002-routing-processor-proposal/poc/semantic_cache/server/requirements.txt
grpcio>=1.59.0
grpcio-tools>=1.59.0
numpy>=1.24.0
pymilvus>=2.3.0
sentence-transformers>=2.2.0
protobuf>=4.24.0
faiss-cpu>=1.7.4  # faiss-gpu